Create `BufferedUdpMetricSink` that buffers metric lines before sending
New MetricSink implementation that buffers metrics sent to it up to a certain
size (user configurable) before sending via a UDP socket. Each metric is written
to a shared (mutex guarded) buffer along with a trailing newline (\n). When the
buffer cannot fit the next metric, it will be flushed.

Metrics that are too large for the buffer (unlikely but possible) will be written
directly to the underlying socket.

Current limitations:

* Mutex poisoning is not handled. It has the potential to panic the thread.
* Metrics are still written to the socket in the calling thread.

Fixes #18