Remove the generic type from StatsdClient
Remove the generic type `<T: MetricSink>` from the StatsdClient to make it more
ergonomic to use.

Prior to this, the easiest way to use the client was to put it behind a pointer
of some kind and reference it like `Box<MetricClient>` or
`Arc<MetricClient + Sync + Send>`. This is still possible but is now not required.
Users of the client may simply refer to it directly without needing to include
the concrete type of the `MetricSink`: `fn foo(some_param: StatsdClient) {}`

This commit also introduces some documentation for the `StatsdClient` that
describes the various ways this client can be shared between multiple threads.

Fixes #45