Include line ending in space required calculation
Include the trailing line ending when determining if there is
enough space in the buffer to write the given byte array. When
the line ending was not included, a write would be attempted
without sufficient space in the buffer which would cause an
excessive number of writes to the underlying `Write` impl.

Fixes #59