Optionally send metrics from MetricBuilder quietly (no result)
Change the behavior of the `MetricBuilder::send()` method to not return
a result based on the sending of the metric. Users that want the result
can now use the `MetricBuilder::try_send()` method for sending
metrics.

On success, the output from sending the metric with `.send()` is
discarded. On error, a user supplied error handler is invoked to consume
the error. By default, if no handler is supplied, the errors are
silently discarded.

The motivation for discarding errors has a few parts:
* We don't want to add another dependency to Cadence.
* There have been issues opened in the past to remove logging from
  Cadence.
* Users that want to do something more elaborate will have a better
  idea of the right course of action than us.

Fixes #65