Optional full Datadog compatibility.
First of all, I'm sorry for opening this huge PR xD. Please don't
be scared (I hate long PRs) I promise is not as bad as it looks ;)
It's mostly docs and tests.I am working on a prototype at work and
we use Datadog extensively,during my research for a Stastd client
I've found this awesome crate and I loved it! But it did not support
(understandably) Datadog specific extensions, so after a bit of
consideration, I decided that instead of going and creating my
own client wrapping by Cadence, I could implement a backwards
compatible, optional extension, that would turn Cadence in a
fully fledged Datadog client when requested.

This PR implements a proposal of such functionality.

- I talked to Datadog guys about it and they were thrilled with it,
they even sponsored me an account. So if you decide to move forward with this,
we can iterate until reaches a mergeable state. Then, we can list cadence
in the Datadog page as the first fully compliant Rust client.
- I'm still missing Datadog "Service Checks" functionality but I
consider I have something worthy to show and I just wanted to see your
opinion on it folks.
- I'm not really sure if the `size_hint` for events makes sense, could you
give it a detailed look?
- There is still a bunch of duplication between `MetricBuilder` and
`EventBuilder` I guess we can fix that before we merge.

Without further ado, the serious stuff:

Implemented a set of extensions under the optional `datadog-extensions`
Cargo feature that, when enabled, will turn Cadence into a fully
fledged Datadog client.

All the extensions were based on the implementation of the official DogStatsd client.

- Distributions: (Also called Global Percentiles): Very similar to histograms
but with the key difference that they histograms percentiles are calculated
server-side (by sending sketches of the histogram data per host to the server),
instead of at the host, making it possible to compute accurate DogstatsD
percentiles across hosts.
- Sets: Which are not a Datadog only feature but are not currently supported.
Sets allow counting unique occurrences of events between flushes, using a Set
to store all occurring events.
- Events: Events to be sent to the Datadog event stream. Useful to report
things like exceptions happening in your system along with their stack-traces.

All of the implementation, when possible, was added inside a `datadog` module
compiled only when the crate is required with the `datadog-extensions` enabled.
Changes in the current outer modules where mostly crate-level visibility tweaks
in order to reuse as much code as possible.

For supporting Sets and Distributions two new traits are added `Setted` and
`Distributed` (following the naming conventions observed in the code, we can
change those names in case we don't like them). Both traits follow the same
conventions seen in the rest of the traits implemented by `StatsdClient`, as in
they have a `set` and `distribution` methods with their respective `*_with_tags`
variation that returns a builder.

In order to ease the usage of all the features including the Datadog extensions,
a new `DatadogMetricClient` is added to make possible bring all the functionality
pulling a single trait. Very much as `MetricClient` does.

In order to add support for datadog events and in the same vein of `Setted`
and `Distributed` an `Evented` trait is added exposing two methods `event`
and `custom_event` which returns a builder. Rationale behind this name is
that events can be customized mush more than just adding tags so the
`*_with_tags` was not a suitable anymore, again, id we decide to change
for something better, I'm up for that too.

I've written as much tests as I could, and to be honest, coming from Ruby: I
think I exceeded a bit xD. Anyways, all the tests follow the conventions found
in the current implementation.

Respective documentation is also added.