Switch to an AtomicBool for QueuingMetricSink stopped flag
Use an atomic bool instead of a mutex + condvar for setting the stopped
flag of the QueuingMetricSink. This avoids combining a mutex and atomics
(since clippy warns about this) and makes the code a bit simpler. The
cost of this is that the stop_and_wait method now contains a spin
loop but this should be fine since it is only used for unit tests.