Replace uses of crossbeam MsQueue with crossbeam_channel
Replace all uses of crossbeam MsQueue (from 0.3 series) with senders
and receivers from the new crossbeam_channel crate. Use an unbounded
channel to keep the same behavior as MsQueue.

Keep the idea of using "posion pill" messages to allow the sink and
any workers or threads to be easily stopped when the sink is dropped.
Additionally, this makes testing easier.

Fixes #79