Don't write a trailing newline when bypassing buffer
Don't write a trailing newline when bypassing the buffered
`Write` implementation. The point of the newline is to separate
multiple values sent as part of a single underlying `.write(buf)`
call. It doesn't make sense to add one when only a single piece
of data will be sent via the `.write(buf)` call.

Fixes #87