Add an option to create a QueuingMetricSink with a bounded queue
Add a constructor to create a QueuingMetricSink with a bounded queue
size to enable back pressure when the wrapped sink cannot keep up with
the rate of metrics submitted to it. The unbounded constructor is still
available for users that want the old behavior.

When the max queue size is reached, writes to the queuing sink will fail
until the wrapped sink begins to drain entries. There is no "default" size
for the bounded queue and so users must make this determination on their
own for now. Based on the expected rate of metrics being sent, a queue
size of 64K entries or more may make sense. This is left as an exercise
for the savvy user.

Fixes #95