Histogram methods to support recording Durations as nanoseconds
Add methods to the Histogram trait to support writing stdlib Durations
as nanoseconds. Histogram values are extension to Statsd and may not
be supported on all servers. Where they are supported, this makes it easier
to record sub-millisecond times.

Note that a u64 (what histograms use) can only represent Durations of
roughly 500 years as nanoseconds. This shouldn't be an issue in practice
but if a Duration longer than this (that cannot be represented as a `u64`)
is passed to the methods in question, an error will be returned.

Fixes #98