Make client traits generic on value types
Change the way each of the metric type traits that make up StatsdClient
are defined so that they are generic on the type of the value being
submitted for the metric. Instead they accept any type that implements
a `To*Value` trait, which each type of metric has. This allows a us to
use a single method for u64 and Duration for timers and accept f64 for
gauges without needing to pollute the corresponding traits with type-
specific methods.

This change also makes any traits that are not intended to be implemented
outside of Cadence sealed, preventing this so that we may make changes
to these traits without fear of breaking consumers of the library.

Fixes #132