Add macros for each metric type that use a global client instance
Add macros for each metric type (`statsd_count`, `statsd_time`, etc.)
that use a global default client instance. Methods are provided for
callers to set the default client after configuring it (the same way
they would normally).

As part of this, the project has been split into two crates. The `cadence`
crate will continue to include the client and all relevant types and traits.
The `cadence-macros` crate will pull in the base crate as dependency and
build on it to expose macros and a default global client.

Fixes #114