Rewrite spy metric sinks to use channels instead of shared writers
Rewrite `SpyMetricSink` and `BufferedSpyMetricSink` to use channels to
share metrics emitted with callers instead of a shared `Write` impl
protected by a `Mutex`. This avoids poisoning the mutex if a test
assertion fails while the mutex is held (which is common for something
meant to be used for testing).

Examples of how tests can be written against the new spy sinks can
be found in the cadence/examples/spy-sink.rs example as well as the
integration tests for cadence-macros.

Fixes #124