Disable criterion benchmarks
At the time it was added, Criterion was the only way to benchmark on
stable Rust. While useful, the number of dependencies it pulls it makes
it very difficult to support older versions of Rust.

Disable benchmarks until such time as our MSRV works with the various
dependencies Criterion pulls in.