Add replace_writer method to MultiLineWriter
The motivation for this change is to implement connection-oriented,
stateful sinks (Virtio/vsock, or TCP).

To enable connection recovery with `MultiLineWriter`, there is a need
to replace the underlying writer. This interface also exposes the
state of the replaced underlying writer using `BufWriter::into_parts()`,
which allows for the potential of more precise recovery.

Test coverage is included.